requirements1.txt
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
click==7.1.2
Django==3.0
django-cors-headers==3.7.0
django-extensions==3.1.1
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Faker==8.2.0
localStoragePy==0.2.3
mypy-extensions==0.4.3
mysqlclient==2.0.3
pathspec==0.8.1
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
